DEBUG unit test availablecoins_tests fails intermittently
this is a debugging commit the previously introduced
availablecoins_tests in the merge of https://github.com/bitcoin/bitcoin/pull/24584

about 30% of the time the test fails with "invalid schnorr signature"
when trying to verify the signature of the transaction it has just
created

this only occurs when both bech32 and bech32m tests are run in the test.
removing either one does not show the issue. even stranger, the issue
only occurs when bech32m is run before bech32 as it is in the upstream
PR. swapping these around so that bech32 is run before bech32m works
around the issue - BUT WHY