test: fix wallet_migration.py
when specifying inputs for sendall, fetching the value of the input
incorrectly used the input loop index instead of the actual prevout
index.

this meant that *sometimes* the sendall call would use the wrong input
resulting in a tx that didn't balance - when the loop index was
different to the vout.