fix: wallet_elements_regression_fundrawtransaction.py
This issue was originally found in an call to fundrawtransaction where a
raw transaction was created with 2 pre-selected inputs, one policy asset
and issued asset, with only a policy asset output specified. This would
result in the call returning "insufficient funds" since it was trying to
select coins for a negative value of the issued asset.

The fix here is to check that the raw transaction contains an output for
each asset type or return an appropriate error message.