rpc: fix rpc typechecking errors
Fixes (most?) of the rpc typechecking errors, at least the ones that are
tested by the functional tests.

However, I've left the rpc typechecking off after fixing these, since
I'm concerned that there are conditional cases that are untested in our
functional tests - and I'd rather not have RPC calls failing in a
release just because the documentation is not correct.